testing rounding method
this txtfile is a list of tuples, each containing the filename, the list of rounded unix ts, and the list of senders in that file. in theory, files containing the same rounded unix ts should have the same senders, if this method works. this is only a list; i have yet to figure out a way to get python to go in and just tell me whether duplicates of a given unix ts has the same sender. this is partly because the unit of analysis is different--email vs. file page.